Implemented add/update/delete and toggle user's admin status functionalities for admin users.

+++ src/main/java/ie/lukeandella/wedding/pojos/Role.java
@@ -9,7 +9,7 @@ public class Role {
     @Id
     @Column(name = "role_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String name;
 
@@ -19,7 +19,7 @@ public class Role {
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
